+++ src/BUPSystem/Revenue budgeting/Revenue budgeting via product.xaml.cs
@@ -44,11 +44,14 @@ namespace BUPSystem.Revenue_budgeting
             set { RevenueManagement.Instance.CurrentFinancialIncomeYear = value; }
         }
 
+        private ObservableCollection<FinancialIncome> NewFinancialIncomeList { get; set; }
+
         /// <summary>
         /// Constructor
         /// </summary>
         public RevenueBudgetingViaProduct()
         {
+            NewFinancialIncomeList = new ObservableCollection<FinancialIncome>();
             CurrentFinancialIncomeYear = RevenueManagement.Instance.CreateFinancialIncomeYear();
             InitializeComponent();
             dgIncomeCustomer.IsEnabled = false;
@@ -127,9 +130,13 @@ namespace BUPSystem.Revenue_budgeting
         {
             try
             {
-            FinancialIncomeList = RevenueManagement.Instance.RemoveEmptyProductIncomes();
-            RevenueManagement.Instance.UpdateFinancialIncome();
-            MessageBox.Show("Intäktsbudgeteringen är nu sparad");
+                if (NewFinancialIncomeList.Any())
+                    foreach (FinancialIncome fi in NewFinancialIncomeList)
+                        RevenueManagement.Instance.AddIncome(fi);
+
+                FinancialIncomeList = RevenueManagement.Instance.RemoveEmptyProductIncomes();
+                RevenueManagement.Instance.UpdateFinancialIncome();
+                MessageBox.Show("Intäktsbudgeteringen är nu sparad");
             }
             catch
             {
@@ -144,14 +151,18 @@ namespace BUPSystem.Revenue_budgeting
         /// <param name="e"></param>
         private void btnDelete_Click(object sender, RoutedEventArgs e)
         {
-            try
+            var result = MessageBox.Show("Är du säker på att du vill ta bort denna raden helt?" + Environment.NewLine + "Detta går inte att ångra.", "Ta Bort Rad", MessageBoxButton.YesNo);
+            if (result == MessageBoxResult.Yes)
             {
-                FinancialIncome fi = (FinancialIncome)dgIncomeCustomer.SelectedItem;
-                RevenueManagement.Instance.DeleteFinancialIncome(fi);
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message);
+                try
+                {
+                    FinancialIncome fi = (FinancialIncome)dgIncomeCustomer.SelectedItem;
+                    RevenueManagement.Instance.DeleteFinancialIncome(fi);
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.Message);
+                }
             }
         }
 
@@ -245,7 +256,7 @@ namespace BUPSystem.Revenue_budgeting
             FinancialIncome obj = e.NewItem as FinancialIncome;
             if (obj != null)
             {
-                RevenueManagement.Instance.AddIncome(obj);
+                NewFinancialIncomeList.Add(obj);
                 obj.ProductID = SelectedProduct.ProductID;
             }
         }
